requirements.txt
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
exceptiongroup==1.0.4
idna==3.4
iniconfig==1.1.1
packaging==22.0
pluggy==1.0.0
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.13
